+++ examples/vue/src/main.ts
@@ -1,6 +1,6 @@
 import { createApp } from "vue";
 import App from "./App.vue";
-import { DemoGhostPlugin } from "@demoghost/vue";
+import { DemoGhostPlugin } from "@demoghostjs/vue";
 import "demoghost/css";
 
 const app = createApp(App);